+++ application/datamapper/cache/filecache.php
@@ -156,13 +156,16 @@ class DMZ_FileCache implements DMZ_CacheInterface
 			return false;
 		}
 		
+		$success = true;
 		foreach ($files as $file) {
 			if (is_file($file)) {
-				@unlink($file);
+				if (!@unlink($file)) {
+					$success = false;
+				}
 			}
 		}
 		
-		return true;
+		return $success;
 	}
 	
 	/**